move exportResultsClicked back to DetailsViewActionMessageCreator
