+++ src/tests/unit/tests/DetailsView/actions/assessment-action-message-creator.test.ts
@@ -65,38 +65,6 @@ describe('AssessmentActionMessageCreatorTest', () => {
         );
     });
 
-    test('exportResultsClicked', () => {
-        const serviceKey = 'html';
-        const event = eventStubFactory.createMouseClickEvent() as any;
-
-        const telemetry: ExportResultsTelemetryData = {
-            source: TelemetryEventSource.DetailsView,
-            triggeredBy: 'mouseclick',
-            exportResultsService: 'html',
-            exportResultsType: 'export result type',
-        };
-
-        const exportResultsType = 'Assessment';
-
-        telemetryFactoryMock
-            .setup(tf =>
-                tf.forExportedResults(
-                    exportResultsType,
-                    serviceKey,
-                    event,
-                    TelemetryEventSource.DetailsView,
-                ),
-            )
-            .returns(() => telemetry);
-
-        testSubject.exportResultsClicked(exportResultsType, serviceKey, event);
-
-        dispatcherMock.verify(
-            dispatcher => dispatcher.sendTelemetry(EXPORT_RESULTS, telemetry),
-            Times.once(),
-        );
-    });
-
     test('selectRequirement', () => {
         const view = VisualizationType.Headings;
         const selectedRequirement = HeadingsTestStep.headingFunction;
